Fix another incompatible character encoding error
In fae72e5 we introduced a fallback for incompatible character
encodings, but that only covered the case where Mapi::Mime#is_multipart?
returns true.

This is a slight variation for the other branch of the conditional.

The error is being raised where we have a `@body` in a different
encoding to the default string encoding. In this case, just always force
the string to be the same encoding as `@body` so we don’t get an error.

This passes the spec for the particular file we’re seeing problems with,
but I’m not at all sure this is the right thing to do. Maybe we should
always treat anything in mapi as ASCII, and then force to UTF-8 for
display? [1]

Fixes https://github.com/mysociety/alaveteli/issues/5783

[1] https://github.com/aquasync/ruby-msg/issues/9#issuecomment-54589744